Remove random change of city name
For whatever reason, we changed the name of cities when the cities associated with an empire were updated. Don't ask me why. I can only speculate that I considered changing the name once a city state fell to another culture with a different language. See St. Petersburg, Leningrad, etc. which were all renamed when they changed hands.

We'll remove this for now.