Performance boost: Distance field reuse / delta updates
Introducing the ability to update distance fields as new seed points are introduced, which gives us a pretty massive performance boost.

Right now, I've added this to the city placement algorithm, but it might work in other places too.

Further, we could cache the distance fields (in BaseObject) and keep updating them as we go.